feat: refactor ThinkingReasoning component and add useThinkingReasoning hook for improved state management
